+++ src/java/org/apache/hadoop/hdfs/server/namenode/BackupStorage.java
@@ -43,6 +43,7 @@ public class BackupStorage extends FSImage {
 
   /** Backup input stream for loading edits into memory */
   private EditLogBackupInputStream backupInputStream;
+
   /** Is journal spooling in progress */
   volatile JSpoolState jsState;
 
@@ -214,7 +215,8 @@ public class BackupStorage extends FSImage {
           waitSpoolEnd();
           // update NameSpace in memory
           backupInputStream.setBytes(data);
-          editLog.loadEditRecords(getLayoutVersion(),
+          FSEditLogLoader logLoader = new FSEditLogLoader(namesystem);
+          logLoader.loadEditRecords(getLayoutVersion(),
                     backupInputStream.getDataInputStream(), true);
           getFSNamesystem().dir.updateCountForINodeWithQuota(); // inefficient!
           break;
@@ -334,11 +336,12 @@ public class BackupStorage extends FSImage {
       // load edits.new
       EditLogFileInputStream edits = new EditLogFileInputStream(jSpoolFile);
       DataInputStream in = edits.getDataInputStream();
-      numEdits += editLog.loadFSEdits(in, false);
-  
+      FSEditLogLoader logLoader = new FSEditLogLoader(namesystem);
+      numEdits += logLoader.loadFSEdits(in, false);
+
       // first time reached the end of spool
       jsState = JSpoolState.WAIT;
-      numEdits += editLog.loadEditRecords(getLayoutVersion(), in, true);
+      numEdits += logLoader.loadEditRecords(getLayoutVersion(), in, true);
       getFSNamesystem().dir.updateCountForINodeWithQuota();
       edits.close();
     }
